+++ src/index.ts
@@ -1,3 +1,10 @@
-export default function parse() {
-  throw new Error('Not yet implemented');
+import changelogParser from 'changelog-parser';
+
+export default async function parse(content: string) {
+  const parsed = await changelogParser({
+    text: content,
+  });
+  return {
+    versions: parsed.versions,
+  };
 }