+++ tests/data.ts
@@ -1,0 +1,35 @@
+const basicChangelog = `
+# changelog
+
+## 0.0.3 - 2023-01-03
+
+### Added
+
+- A3
+
+### Removed
+
+- B3
+
+## 0.0.2 - 2023-01-02
+
+### Added
+
+- A2
+
+### Removed
+
+- B2
+
+## 0.0.0 - 2023-01-01
+
+### Added
+
+- A
+
+### Removed
+
+- B
+`;
+
+export default basicChangelog;